+++ specs/jsdoc/fixtures/misc.js
@@ -1,6 +1,5 @@
 /**
  * @name publicFunc
- * @bit
  * @public
  */
 function publicFunc() {